fix: refactor leaf_index to u64 everywhere (#360)

+++ contracts/src/interfaces/IWorldIDRegistry.sol
@@ -45,7 +45,7 @@ interface IWorldIDRegistry {
     /**
      * @dev Thrown when a requested leaf index does not exist.
      */
-    error AccountDoesNotExist(uint256 leafIndex);
+    error AccountDoesNotExist(uint64 leafIndex);
 
     /**
      * @dev Thrown when a recovered signature address is the zero address.
@@ -80,7 +80,7 @@ interface IWorldIDRegistry {
     /**
      * @dev Thrown when the leaf index does not match the expected value.
      */
-    error MismatchedLeafIndex(uint256 expectedLeafIndex, uint256 actualLeafIndex);
+    error MismatchedLeafIndex(uint64 expectedLeafIndex, uint64 actualLeafIndex);
 
     /**
      * @dev Thrown when the recovered signature does not match the expected authenticator address.
@@ -95,12 +95,12 @@ interface IWorldIDRegistry {
     /**
      * @dev Thrown when a nonce does not match the expected value.
      */
-    error MismatchedSignatureNonce(uint256 leafIndex, uint256 expectedNonce, uint256 actualNonce);
+    error MismatchedSignatureNonce(uint64 leafIndex, uint256 expectedNonce, uint256 actualNonce);
 
     /**
      * @dev Thrown when a recovery counter does not match the expected value.
      */
-    error MismatchedRecoveryCounter(uint256 leafIndex, uint256 expectedRecoveryCounter, uint256 actualRecoveryCounter);
+    error MismatchedRecoveryCounter(uint64 leafIndex, uint256 expectedRecoveryCounter, uint256 actualRecoveryCounter);
 
     /**
      * @dev Thrown when a pubkey ID overflows its uint32 limit.
@@ -110,7 +110,7 @@ interface IWorldIDRegistry {
     /**
      * @dev Thrown when a recovery address is not set for an account.
      */
-    error RecoveryAddressNotSet(uint256 leafIndex);
+    error RecoveryAddressNotSet(uint64 leafIndex);
 
     /**
      * @dev Thrown when an authenticator does not exist.
@@ -120,7 +120,7 @@ interface IWorldIDRegistry {
     /**
      * @dev Thrown when an authenticator does not belong to the specified account.
      */
-    error AuthenticatorDoesNotBelongToAccount(uint256 expectedLeafIndex, uint256 actualLeafIndex);
+    error AuthenticatorDoesNotBelongToAccount(uint64 expectedLeafIndex, uint64 actualLeafIndex);
 
     /**
      * @dev Thrown when trying to update max authenticators beyond the natural limit.
@@ -150,7 +150,7 @@ interface IWorldIDRegistry {
      * @param offchainSignerCommitment The offchain signer commitment for the account.
      */
     event AccountCreated(
-        uint256 indexed leafIndex,
+        uint64 indexed leafIndex,
         address indexed recoveryAddress,
         address[] authenticatorAddresses,
         uint256[] authenticatorPubkeys,
@@ -168,7 +168,7 @@ interface IWorldIDRegistry {
      * @param newOffchainSignerCommitment The new offchain signer commitment.
      */
     event AccountUpdated(
-        uint256 indexed leafIndex,
+        uint64 indexed leafIndex,
         uint32 pubkeyId,
         uint256 newAuthenticatorPubkey,
         address indexed oldAuthenticatorAddress,
@@ -186,7 +186,7 @@ interface IWorldIDRegistry {
      * @param newOffchainSignerCommitment The new offchain signer commitment.
      */
     event AccountRecovered(
-        uint256 indexed leafIndex,
+        uint64 indexed leafIndex,
         address indexed newAuthenticatorAddress,
         uint256 indexed newAuthenticatorPubkey,
         uint256 oldOffchainSignerCommitment,
@@ -200,7 +200,7 @@ interface IWorldIDRegistry {
      * @param newRecoveryAddress The new recovery address.
      */
     event RecoveryAddressUpdated(
-        uint256 indexed leafIndex, address indexed oldRecoveryAddress, address indexed newRecoveryAddress
+        uint64 indexed leafIndex, address indexed oldRecoveryAddress, address indexed newRecoveryAddress
     );
 
     /**
@@ -213,7 +213,7 @@ interface IWorldIDRegistry {
      * @param newOffchainSignerCommitment The new offchain signer commitment.
      */
     event AuthenticatorInserted(
-        uint256 indexed leafIndex,
+        uint64 indexed leafIndex,
         uint32 pubkeyId,
         address indexed authenticatorAddress,
         uint256 indexed newAuthenticatorPubkey,
@@ -231,7 +231,7 @@ interface IWorldIDRegistry {
      * @param newOffchainSignerCommitment The new offchain signer commitment.
      */
     event AuthenticatorRemoved(
-        uint256 indexed leafIndex,
+        uint64 indexed leafIndex,
         uint32 pubkeyId,
         address indexed authenticatorAddress,
         uint256 indexed authenticatorPubkey,
@@ -304,7 +304,7 @@ interface IWorldIDRegistry {
      * @param nonce The signature nonce for replay protection.
      */
     function updateAuthenticator(
-        uint256 leafIndex,
+        uint64 leafIndex,
         address oldAuthenticatorAddress,
         address newAuthenticatorAddress,
         uint32 pubkeyId,
@@ -329,7 +329,7 @@ interface IWorldIDRegistry {
      * @param nonce The signature nonce for replay protection.
      */
     function insertAuthenticator(
-        uint256 leafIndex,
+        uint64 leafIndex,
         address newAuthenticatorAddress,
         uint32 pubkeyId,
         uint256 newAuthenticatorPubkey,
@@ -353,7 +353,7 @@ interface IWorldIDRegistry {
      * @param nonce The signature nonce for replay protection.
      */
     function removeAuthenticator(
-        uint256 leafIndex,
+        uint64 leafIndex,
         address authenticatorAddress,
         uint32 pubkeyId,
         uint256 authenticatorPubkey,
@@ -376,7 +376,7 @@ interface IWorldIDRegistry {
      * @param nonce The signature nonce for replay protection.
      */
     function recoverAccount(
-        uint256 leafIndex,
+        uint64 leafIndex,
         address newAuthenticatorAddress,
         uint256 newAuthenticatorPubkey,
         uint256 oldOffchainSignerCommitment,
@@ -393,7 +393,7 @@ interface IWorldIDRegistry {
      * @param signature The signature from an existing authenticator authorizing the update.
      * @param nonce The signature nonce for replay protection.
      */
-    function updateRecoveryAddress(uint256 leafIndex, address newRecoveryAddress, bytes memory signature, uint256 nonce)
+    function updateRecoveryAddress(uint64 leafIndex, address newRecoveryAddress, bytes memory signature, uint256 nonce)
         external;
 
     ////////////////////////////////////////////////////////////
@@ -414,7 +414,7 @@ interface IWorldIDRegistry {
      * @dev Returns the recovery address for the given World ID (based on its leaf index).
      * @param leafIndex The index of the leaf.
      */
-    function getRecoveryAddress(uint256 leafIndex) external view returns (address);
+    function getRecoveryAddress(uint64 leafIndex) external view returns (address);
 
     /**
      * @dev Checks whether `root` is known and not expired according to `rootValidityWindow`.
@@ -431,18 +431,18 @@ interface IWorldIDRegistry {
      * @dev Returns the signature nonce for a leaf index.
      * @param leafIndex The leaf index to query.
      */
-    function getSignatureNonce(uint256 leafIndex) external view returns (uint256);
+    function getSignatureNonce(uint64 leafIndex) external view returns (uint256);
 
     /**
      * @dev Returns the recovery counter for a leaf index.
      * @param leafIndex The leaf index to query.
      */
-    function getRecoveryCounter(uint256 leafIndex) external view returns (uint256);
+    function getRecoveryCounter(uint64 leafIndex) external view returns (uint256);
 
     /**
      * @dev Returns the next available leaf index.
      */
-    function getNextLeafIndex() external view returns (uint256);
+    function getNextLeafIndex() external view returns (uint64);
 
     /**
      * @dev Returns the depth of the Merkle tree.
